[CHORE] 멤버 정보 view 컴포넌트 분리 (#293)
